Create change
Patch-set: 1
Change-id: Ifb8bba079bd7b778e67715e728f258c6a00037a1
Subject: Abuse filtering is now correctly detecting a disallow/warning; the previous version only looked for the existance of a message, which may or may not be the right thing to do depending on the consequences checked.  We need access to the names of the actions taken, not just an already-translated message.
Branch: refs/heads/master
Commit: de942fe420219c8c6361f520b3783b7c14b75056