Update patch set 1
Patch Set 1:

Siebrand - could you please explain how to fix the issue? I can't find examples of this in the docs or in the core code.

I have a timestamp from the database I need to display - what do I pass to the message formatter? Do I need to use a special method to assign parameters? How should the message format show that this is a time and date (timestamp)?  Do you have an example of code that does this I can look at?  

Thanks!

Patch-set: 1