Send the bucketing version as part of the key
Gets around a bug in mw.user.bucket causing users with a cookie not to be
re-bucketed when the version changes.

Also removed the redundant and confusing AFT cookie, after confirming that it
was not the source of the problem and simply duplicated what mw.user.bucket was
already doing.

Change-Id: I4b8281df650e35ee7ba6d7a31bf99bde7163d102