Update patch set 7
Patch Set 7:

I didn't check the last patch set, but I don't see special handling for [[Special:Contribs/newbie]] in contributionsLineEnding for specifying the username.

I did see a problem with it in the previous review.
Please check that in [[Special:Contribs/newbie]] it shows the user who wrote the feedback - if it shows the username you can ignore this comment :)

Patch-set: 7