Update patch set 1
Patch Set 1: There's a problem with this change, please improve

I'm not certain all ".errorbox" errors could conclude that the actual article is not visible. It could be a future minor error that affects some minor part of the page. Perhaps we should actually look if there is an article in the DOM somewhere.

Patch-set: 1