Update patch set 3
Patch Set 3:

WTF?!? It's mind-boggling to me that shortening an existing index would solve anything. This can't be the real cause of the issue. Was the index present in the .sql file but not deployed on the cluster? In that case we don't need a software change, we just need to deploy the index. Was MySQL being stupid and not using the index? In that case we don't need a different index, we need a FORCE INDEX.

Patch-set: 3