Update patch set 3
Patch Set 3: There's a problem with this change, please improve

I reviewed the index structure in this change as if it were a new index as the original version was never deployed to prod and the .sql file with the actual ALTER is new.  Either version addresses the performance of the query that impacted production but the original version should be superior as the table grows. 

The original version is now deployed to prod.  I'd like to see a revised index_user_data.sql file committed for the actual three column index structure  instead of having this change completely abandoned.

Patch-set: 3