Update patch set 1
Patch Set 1: There's a problem with this change, please improve

I think this localises the link target, too, while it shouldn't. Is that correct? I.e. the text should be localized, the URL should be taken from the content language. Have a look at how this is done in the Narayam and WebFonts extension. IIRC it's done by assigning the URL to a global, which can later be used in JS.

Patch-set: 1