Update patch set 3
Patch Set 3:

The latest patch set merges the rules for both east and west arrow and relies on CSSJanus for flipping.

Note: For testing this patch set, it is recommended to apply change I5d24c7d8 to core. It fixes a problem with the box shadow in RTL environment. As I write this comment, it's not merged yet.

Patch-set: 3