Update patch set 51
Patch Set 51: There's a problem with this change, please improve

(7 inline comments)

It's very possible I do not understand how the select wrapper works, but it looks like the blob comment table is being joined against for every fetch.

If that is the case, then it should be rolled into one table.

Patch-set: 51