Update patch set 54
Patch Set 54:

The blobs table request was made when all aft tables were going into the monolithic unsharded enwiki db, with the thought of possibly later moving them elsewhere. Since aft will now get a separate shard, it isn't needed an longer. 

I've heard conflicting reports of where we stand with the DataModel class as it relates to this change. I personally think aft at this time should use the standard mediawiki data access methods, which already support extensions using their own shard. Aaron has demonstrated this in a few lines of code. If we are to get behind a new DAO layer (or alternate paterns) with integrated caching and potentially sharding, I'd like it to be in core with full input and review from platform.

Patch-set: 54