Update patch set 54
Patch Set 54: There's a problem with this change, please improve

Let me phrase the blob question differently.

I get that there are sometimes good reasons to separate blob and text types form the rest of the data even when there is a 1:1 relationship.

In this case there will not be a 1:1 relationship. Only a small percentage of ratings have a long comment.  That seems like an expensive join for data that only sometimes exists and is only sometimes needed for display even when it exists.

Without benchmarking I can't tell you if not joining then having to do a second query in the 3% of cases where it exists is faster then joining every time and getting 97% nulls.

But, can't the application logic support a summary fetch for where the long comment is not going to be displayed and a full fetch for the times that the long comment is needed?  That way you can skip the join most of the time.

Patch-set: 54