Update patch set 28
Patch Set 28: Code-Review-1

The behavior for interacting with your own feedback is confusing. The 'Hide my post' link doesn't make it clear if you are just hiding your post from yourself or from everyone, and it certainly doesn't make it clear that you are actually flagging your own post as inappropriate. And if you have permission to see inappropriate feedback, it doesn't actually hide it at all; it just changes the status. Instead of having a separate interface for reviewing your own posts, why not just use the regular reviewing interface, but limit it to 'Mark as Inappropriate'? That way it is clear what action you are taking and the interface is consistent.

Patch-set: 28