Update patch set 28
Patch Set 28:

(1 comment)

Kaldari:
There are users (not-confirmed users) who do not have sufficient permissions to the toolbox-actions (on the right - mark as useful, resolve, non-actionable, inappropriate). Their tools (mark as helpful/unhelpful, flag as inappropriate) are not in the toolbox, but below the post.
To them, it would be confusing to suddenly, for some (their own) feedback find something in the toolbox to mark it as inappropriate. The wording "hide my own post" is because they're not familiar with the other toolset. For consistency, this same behavior is currently applied to editors who actually do have permissions to the toolbox-actions.

You do have a point though. A better solution would probably be to add the action in the toolbox again, for those who do have permissions to see that toolset (it once was this way). It's probably still not the most optimal UX though, as it's rather inconsistent. I'll loop in Pau as well, so we can perhaps figure out a better UX solution later.

Patch-set: 28