Update patch set 1
Patch Set 1:

As I understand it, the "only" shouldn't make any difference, since "@media print" should only be applied to print as well.
For some reason, that did not really seem to be the case.
I've just tested the original code (= not this patchset) again (since I was going to try out with only), but this does not even seem to be the case anymore - everything's working fine without this patch now as well.
I'm going to abandon this for now and consider it to have been an issue with the CSS minifier that has since been fixed ;)

Patch-set: 1