Update patch set 2
Patch Set 2: Code-Review-2

This seems like a terrible idea. Why would we want to make it easier for people to post without comments? Posts without comments have no value to editors. They just create more work. I realize this is a design issue and not a coding issue, but as a Wikipedia editor I feel like this is a very bad idea.

Patch-set: 2