Update patch set 3
Patch Set 3: Code-Review-1

aft_id is (or should now) be using a time sorted uuid algorithm, it's the primary key, and the primary key is always usable for an ORDER BY even when it isn't in a secondary key.  So instead of ordering by aft_timestamp and adding it to keys, can you order by aft_id instead?

Patch-set: 3