Left-pad legacy aft_id values with null-bytes, instead of right-padding
aft_id values actually increment over time, although because the legacy values are
currently being right-padded with null-bytes, they can not be ORDERed BY that way.
This patch will clean this up: left-padding them with null-bytes will fix this and
make it possible to perhaps some day sort on aft_id for a time-based sort.

Change-Id: I2e9b0e2014b0f4a70c87c39817cd6d955de020cb