Update patch set 2
Patch Set 2:

The config files of wmf are okay, because there are skipping a permission, if the group is unknown (fixed with I57fe2d85), but the default permissions shipped with the extension are not okay, because there are setting the permissions also or non-existing groups. That is fixed by this patch set.

Patch-set: 2