Alternative method to detect the existence of a user/talk page discussion
The original method of getting sections from parser output was quite expensive
when parser output is not in cache.
Instead, after clicking the link to add something to a user/talk page, I'll
pass on some parameters, which will be read after saving the page & a flag will
be added in the AFT table to indicate an entry was discussed about.

The slight disadvantage is that at a later point, someone may remove the section
from the talk page, making it no longer "discussed" (but AFT's table will still
think it's discussed). This does not really matter too much; the current
solution was not even capable to recognize section titles that are changed from
what the tool suggests ;)

Change-Id: I84677fedddf09b31d95bcc6aedac6c4fad4a50f6