(bug 37427) Filtering by "helpfulness", ranking mistake
Changed "helpful" filter to be based on the percentage of helpful/unhelpful, not the absolute difference.

E.g.: 10x helpful & 4x unhelpful (10 - 4 = 6) would have outranked 5x helpful & 0x unhelpful (5 - 0 = 5)
This was especially confusing since the helpfulness is displayed to users as a percentage, which is not what the sort was actually based on.

Bug: 37427
Change-Id: I611f2c73244f97e80956cfbf07dbb548df17bc58