Update patch set 3
Patch Set 3: -Code-Review

I'm taking off the -2, but I don't really understand why we need an extra UI element for this. If submits with no content get handled differently, why don't we just do that automatically. In other words, detect if the comment is less than X characters (6 maybe?), and if so, allow the comment but hide it from the comment viewing interface since it is an empty or worthless comment.

Patch-set: 3
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0