Update patch set 4
Patch Set 4: Code-Review-1

Bark bark, here comes the third dog. Actually you're both wrong/right.

Matt is correct in realising that mw.Api abstracts both ajax errors and API errors as failures to resolve the promise. However it's signature is different.

It responds with 'code' as first argument (which can be one of many, see mw.Api.errors for a list, I believe there may be even more). The second argument contains extra details. Depending on the error code this can be a string, mediawiki API response data or ajax parameters. Only when the failure is a result of mediawiki data.error will the second argument be the 'data' data this variable used to refer to.

Patch-set: 4