Update patch set 6
Patch Set 6:

I don't know if this is related to the new changes, or just something wonky with the way apis are handled in wm, but when testing page id and title in the same request i get the following response:

{
  error: {
    code: "afvfafvfinvalidparammix",
    info: "The parameters afvftitle, afvfpageid can not be used together"
  }
}


Is afvf supposed to be doubled up in the code like that? Other error messages like a non-existant title get 'afvfmissingtitle' without the afvf doubled up.

Patch-set: 6