Don't show enable/disable links if AFTv5 is status is set via categories
The enable/disable links will enable or disable AFTv5 via page protection
settings, they will not change anything to the categories. Right now, a
"AFTv5 is disabled for this page" status box is being shown for pages where
AFTv5 is not enabled via opt-in, but is via whitelist.
Since opt-in/-out can't edit the page and remove the appropriate category, it
makes no sense to even show the possibility to enable/disable if a white-/
backlist category is applied.

Also: the status box did not check the default value. If no protection
level is explicitly set, it'll determine the default level based on lottery.
For pages that win the lottery, the default is that AFT is enabled, but it
would still show the status box.

Change-Id: Id4d6fe0c8322332d2404724fd60123bf87d3b28e