(bug 49174) Database error when trying to protect a page for creation
AFT's hook ignored title protection.
The page protection hook is also called when attempting to protect non-existing
pages. The existing code would insert an entry with pr_page = 0 (because that's
the id it'll have if the page doesn't exist). If doing this multiple times,
duplicate key errors would be the result.

This will hide the AFT settings & don't process the submitted form if the
article does not yet exist.

Bug: 49174
Change-Id: I77320d6354d688c60f6fb5d2ebe8a42d4b46a3a7