Make the opt-in/-out feature optional per wiki
AFTv5 is in the process of being disabled on dewiki. Existing feedbacks should
still be accessible for now, but it should no longer be possible to submit
new feedback.

Via config, one could already "disable" lottery & whitelist/blacklist categories
but not the ability to opt-in for AFTv5. This patch introduces that.

Setting $wgArticleFeedbackv5EnableProtection to false will get rid of the AFTv5
entry in ?action=protect, will ignore existing opt-in/-outs, and will get rid of
the links to opt-in/-out.

Bug: 50926
Change-Id: I97824072bcd9ead44ddc4e3605a2fd932bda3141