Fix auto-archive script
Auto-archive maintenance script (archiveFeedback.php) currently fails with an
exception thrown by ArticleFeedbackv5Model::validate()
This function checks some of the data being saved and, if incorrect, throws an
exception. One of these checks if looking if the page id is valid.
Apperently, for some old feedback, the article can't be fetched (probably
deleted). While it is great that, before inserting, this data is validated, it
is not so great in this usecase, where it renders us unable to update the entry
(unless we were to assign it to a new page, or delete it)

I've added a switch to make this validation optional. When inserting data, we
should definitely validate, but flagging feedback shouldn't be hindered.

Change-Id: If0d6962aa38e9a93e0cec37db2b0405683295f4a