Purge page cache for every AFTv5 protection change
Visibility of AFTv5 form is (in part) defined via page_restrictions.
This restriction level will be exposed to JS (ConfigVars); purging the cache
when this level changes makes sure the value exposed is always up-to-date.

Cache was already being purged, but only when restriction level was changed via
API call to ApiSetStatusArticleFeedbackv5. The restriction level can also be
changed via the hook in ?action=protect (in which case this purge was not being
performed).

Change-Id: I6bbec60f42a7aa21f691e175b963bab794387055