Adding a space character between links in the special page
I did it in order to avoid bad rendering in languages like Persian

Languages like Persian has different rendering of a character in cases of they
are joined or separated. Browsers consider last character joined instead of 
separated and renders it incorrectly for example you can see the bug.

I tested it in tools.wmflabs.org/wikitest-rtl/w/index.php and it worked
properly for the Persian language.

Bug: 53488
Change-Id: I0346567d9234e5568effcfc6ce9051839ed3ed43