Update patch set 1
Patch Set 1: Verified-1

These entries in `logging` are not only used by WMF's core logging code, but also to fetch a list of all actions performed on feedback entries, in AFTv5.
At the very least, there's other code in AFTv5 that needs to be changed, that currently expects log_title & log_namespace to be the special page:

* ArticleFeedbackv5Activity::buildWhereFeedback & it's callers in ArticleFeedbackv5.activity.php
* maintenance/legacyToShard.php & maintenance/loggingUpdate.php (or even better: a new maintenance script should be created & added to DB updater to migrate existing log entries)

At first sight, looks like this is all that'd need to also be changed in order to make the switch to correctly use the page title (instead of the current Special title).

However, at some point, I remember looking into this and not going through with it as there were other non-trivial problems with it.

I'll check again in more detail soon, as it would definitely be better to use the real page title, if possible!

Patch-set: 1
Reviewer: Matthias Mullie <214@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=-1