Update patch set 3
Patch Set 3:

AFTv5 is no more developed, I wish we could terminate it entirely.

I am not sure why we have rake-jessie for that repo since there is no Rakefile. Given they are unlikely to pass, I think the easiest is just to stop triggering rake-jessie for aftv5 and abandon this change.

Patch-set: 3