[WIP] Cleanup AFTv5 tooltip and convert it to mw-collapsible element
Deleted the tooltip files, reorganized the HTML and CSS for the
tooltip, and removed an unused and unnecessary PSD file named
"images.psd" (propiertary file anyways which can only be accessed
for those who have Photoshop installed, which costs). Instead of
a tooltip now, it's a mw-collapsible element which is collapsed
('mw-collapsed') by default.

Bug: T146524
Change-Id: Ib894f5238e22b17af0997ce15e4f30b73f583b0b