Update patch set 1
Patch Set 1:

I'm sorry, but +/-400 lines is way to much going on in a single patch. This will require a significant amount of review time, and I'm not going to invest this on an entirely unused extension. (But I still care about this extension, that's why I'm writing this.) Please split this patch into three or more, e.g. start with one that does nothing but adding these spaces to "function ()", and end with one that does nothing but removing the "json" format.

Patch-set: 1