Update patch set 1
Patch Set 1:

> By "entirely unused" do you mean "entirely unused *at WMF*"?

Yes, I meant the Wikimedia cluster. I forgot this extension is in use in other places. Thank you for the reminder.

I still think my suggestion is very sensible, and actually extremely simple to do: Just upload a patch that does nothing but adding spaces in "function (", and rebase this patch on top of it. This will reduce this patch by almost 100 lines.

Patch-set: 1
Label: Verified=0