Update patch set 1
Patch Set 1:

> Just upload a patch that does nothing but adding spaces in "function (", and rebase this patch on top of it

That is going to be an insane amount of work for effectively no purpose. The entire rebase would likely need to be done manually, when the only difference in review is a small *number* of easily examined lines, which would all need to be individually looked at regardless of what patch they are in, while also inviting conflicts on two patches as opposed to one.

Thiemo Mättig, while I agree there is often correlation between number of lines changed and difficulty of reviewing a patch, it is not the only thing you should be paying attention to when gauging this, and much depends on *what* is actually being changed as well.

Patch-set: 1