Update patch set 1
Patch Set 1:

I'm sorry you feel this way, but to give the impression that the length here is inherently bad is wrong.

Now if you'd said it might have been better to have started here with only fixing the indentation, that I might even agree with. It's a relatively minor point, given that this is nothing but stylistic stuff anyway, but it would make the rest easier to see.

Patch-set: 1
Label: Verified=0