Update patch set 2
Patch Set 2:

Test failure is T159058 and is essentially unrelated to your code.

However, I'm not sure about the overall justification of this task: some of your changes (e.g. to maintenance/archiveFeedback.php) appear just outright wrong because this extension's internal name is "ArticleFeedbackv5", not "Article Feedback" (even if the latter is the user-friendly name which might be shown in the UI to the end-user).
So I believe you'll want to use $this->requireExtension( 'ArticleFeedbackv5' ); in the maintenance scripts that don't yet have it (such as data/maintenance/DataModelPurgeCache.php) and simply not touch the maintenance scripts which already have that line present (which would be most of them).

Patch-set: 2
Reviewer: Jack Phoenix <243@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0