Use the right prefix for AbuseFilter variables!
In AbuseFilter we recently changed the prefix 'article' to 'page', and
changed everything in our code and related extensions to be compliant.
However, this extension directly calls AbuseFilter::generateTitleVars
with the 'ARTICLE' prefix, which could potentially break AbuseFilters
for this extension. Fortunately, this didn't happen thanks to an extra
parameter added to the function, but the prefix should really be changed
soon!

Bug: T173889
Change-Id: I909a99e80a895a9b009c33ea72e8e0a4ea0a1375