Update patch set 5
Patch Set 5:

@Thiemo: Thanks for the CR again, nice catch!

I agree that in general API i18n seems like a waste of time if and when the strings are never going to be shown to users via "normal" endpoints -- if and when someone's hitting api.php directly with the intention of making bot requests etc. via the API, they're presumably technical enough to be able to access, read and understand [[mw:API]] and the like, even if it's written in English.
That being said, I'm not 100% sure if these strings are indeed going to be shown to users as-is or not, hence why I went with the "let's just i18n them" option. If it turns out that these are unused, it's easier to remove the source strings (and, in time, translations) than to add dozens of translations readily.
Of course my *personal* wish is that people would spend their time translating user-facing strings instead of anything prefixed with apihelp-*. :)

Re:support status, you're right that AFTv5 isn't probably among the most popular extensions out there and while it started out as a WMF project it's no longer deployed on WMF projects, but Brickipedia, the English-language LEGO wiki hosted by ShoutWiki, is still using it and just like in 2016, I'm not aware of any plans to drop it either, hence why I'll have to keep on maintaining it (see T144677 on Phabricator for some details); thus any and all help and code review I can get is very much welcome and appreciated!

Patch-set: 5
Reviewer: Jack Phoenix <243@e9e9afe9-4712-486d-8885-f54b72dd1951>