Update patch set 5
Patch Set 5: Code-Review+1

> […] Brickipedia, the English-language LEGO wiki hosted by ShoutWiki, is still using it and just like in 2016, I'm not aware of any plans to drop it either, hence why I'll have to keep on maintaining it […]

This extension and I, we do have a history. I was one of the more active bug reporters back then when it was running in the German Wikipedia for a short period of time. It makes me happy to hear the extension is still in use somewhere. That's great news. Thanks for letting me know.

Patch-set: 5
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1