Adjust .articleFeedbackv5-panel widths for better responsiveness
Tested under MW 1.33.0 with 33 different MW skins, FOSS and proprietary alike, by loading the Main Page and playing around with it.
Improves the display of the AFTv5 widget for the following FOSS skins:
* DeskMessMirrored
* Dusk
* DuskToDawn
* Gamepress
* MonoBook
* Timeless

While this isn't pixel-perfect at least for MonoBook's and Timeless' responsive modes, this is a huge improvement over the AFTv5 widget overflowing everywhere, as it did prior to this change.

Bug: T34293
Change-Id: Iccba56744126d71a16d99905cc7d059fe6d547cf