Kill all $wgUser fallbacks
This makes all $user parameters non-optional where $wgUser would be used
as fallback.
NOTE: This patch also adds some $wgUser usage. In some places, that
could be easily avoided, but that's for a follow-up patch.

Coincidentally, this also fixes a bug where $wgUser was used even when
not declared as global.

Bug: T241631
Change-Id: I2dcb9529187f9655bc8d0f73a706af3ae43083b2