Avoid deprecated $wgAbuseFilterCustomActionsHandlers and fix up
Even after I51e69736417094cf1ad40a93beea66c713379afe, the AF integration
part was still broken. Firstly because it checked for "AbuseFilter", but
the extension is called "Abuse Filter", so that code was never
triggered. Second, the system user case in Flagging isn't correctly
handled, and it will result in either typehints failing, or some methods
later on screaming that they can only be called when the performer is
not a system user.

As for the deprecated global, it was replaced with the new hook which
allows a higher degree of customization and more structured information
about consequences. Unfortunately, the previous code was a bit hacky,
in that it added the new consequences at runtime, using as callback a
method in the Api module, which we cannot do with the new system. The
hacky solution implemented here is to use a static property instead,
which for this limited use case should be equivalent in terms of
functinoality. The proper solution would look different, and could
involve e.g. another stateful service which keeps track of what flags
should be added or sth like that.

However, given the level of breakage reported above, this quick solution
seemed acceptable in the opinion of the author.

Change-Id: Id1e96e1321d3f781c9796bb177c4e8d094005520