Dramatically minimize .svg files
Turns out a lot of these files contain the entire gradient library
from the drawing tool that was used. Removing all unused code gets
rid of a whopping 1.5 megabytes (about 120 kb when compressed).

This is only a first pass, removing as much code as possible without
making any visible changes. I intentionally don't use SVGO for this
first step. This is for a later patch.

Bug: T337456
Change-Id: If5d674da215614948495d7a6bf163e0d32c07506