Replace missing message with generic error message
As far as I can see this message is shown when $.ajax() fails, which
implies a network error, the server being down, or similar. There is
not much the user can do in this situation other than try again.

"articlefeedbackv5-error" resolves to "An error has occurred. Please
try again later."

I think this is not worth a seperate error. What else should it say?

Bug: T154883
Change-Id: Ibea31b62ed59529ba95ff1ef2f63301f530125bc