Drop a lot of unused images and CSS
A lot of this is unused since Ic99ba85 where the .js code was
removed, but not the .css and not the images.

I made this patch by reading the removed code from Ic99ba85 line by
line and doing a full-text search for every CSS class name and id
that was removed in Ic99ba85. When I couldn't find it any more in
the remaining code I removed it from the .css as well.

This resulted in a lot of images not being referenced from the .css
any more.

A few background images can be replaced with modern CSS. All
officially supported browser support this now.

Bug: T337456
Change-Id: I7b76f1ded78772453852d8db53459faa6038c7cf