Hopefully make the Special:Contributions code a bit clearer and also make phan not complain about XSS vectors anymore
Now the link to the AFTv5 special page gets built in wikitext as opposed to via LinkRenderer and rawParams() is no longer used and whatever <span>s and such $formattedTime may contain are assumed to be wikitext-safe (and it looks like that's the case, based on some quick testing).

Previously briefly unused $5 parameter got repurposed to be the AFTv5 feedback ID (32 characters long, and still not an int), which is used to build the link to the AFTv5 special page in the "articlefeedbackv5-contribs-entry" i18n message.

Change-Id: I9a26f3ce8f3f29640b55d54a74a06e8cb227f0ee