Add a machine-readable output format to articlefeedbackv5-view-feedback API
The default rendering mode is changed from HTML to the more structured, machine-readable version with this patch.
The old HTML mode still exists for backwards compatibility and for use by jquery.articleFeedbackv5.special.js and it is activated by specifying afvfmode=html in the query params for the module.

TODO: So much more testing :)
TODO: I hate the name "ApiArticleFeedbackv5Render" since it's *not* an API module but an utility class used by the main feedback view API class, yet it's appropriate: it's an API-*related* class used by the API module and it's a subclass of ArticleFeedbackv5Render class, the main UI rendering class. Suggestions for a better name are more than welcome.

Bug: T53296
Change-Id: I9053d6b5c078e25dde176014e4d921633a0a8898