Make use of the ?? operator to avoid repetition
Notes:

* The …::$lists array really is structured differently in the base
class and the (only) subclass. In the subclass the conditions are in
a "conditions" sub-element. But not in the base class.

* This patch removes the ?: operator. That operator would also react
on e.g. false, while ?? only reacts on null (as well as when the array
element is missing entirely). I made sure this is safe. Maybe it was
needed for one of the older models (v1…v4), but this code is gone.

Change-Id: I680439d4febf3f0224a433fa27ad11350a6ed8ef